Use spellchecker library to determine real words
Before this commit we would need to reach out to the Webster Dictionary
API to determine if a word was real or not. This added additional
latency to the request for checking and is subject to rate limiting
depending on how their api is implemented.

In order to remove that dependency this commit adds in the use of a
spellchecker library which has a decent dictionary that is local which
can be used to determine if a word is actually a word or not.

This commit also removes the use of a hard-coded private API key which
should be avoided when committing to a public repo.